move login function to utils
